requirements.txt
flask==1.1.1
flask-bootstrap==3.3.7.1
pytest-mock==2.0.0
pyfakefs==4.0.2
plotly==4.5.4
scipy==1.4.1
